// src/services/model/ViewRelationship.ts
/**
 * unknown
 * unknown
 *
 * OpenAPI spec version: last
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

/**
  * @property `[child]` 
  * @property `[id]` id
  * @property `[parent]` 
  */
export interface ViewRelationship {
  "child"?: models.KnowledgeView;
  /**
   * id
   */
  "id"?: number;
  "parent"?: models.KnowledgeView;
}